Configuration: Python extensions have to be manually activated
Until now, the '--with-python' configure option had described itself as
"enable support for python" (which is contradictory to the option
name!). Also, when it was given, a "no" was given to the "with_python"
variable (which is again contradictory!). A third contradiction came when
we checked for its value: we would only attempt to do the Python checks
when 'with_python' was 'no'!

So in the end from the user's perspective, only the documentation of this
option was wrong, not its operation (the Python checks would only occur
when '--with-python' was given).

With this commit, these contradictions have been addressed: the description
has been corrected and the variables are now given a value of 'yes'.