NEWS: preparations for next release
Until now, the top element of the 'NEWS' file was for version 0.20. But
that has been released and in the next series of commits, new things will
be added. I also noticed that updates in the annoncing script required some
changes in 'doc/release-checklist.txt'.

With this commit the new titles of the 'NEWS' file have been added and the
missing steps in the release checklist have been added. Also the notice at
the first page of the PDF book has also been added.