Library (arithmetic): mknoise-poisson using each pixel for scatter
Until now, the 'mknoise-poisson' operator of Arithmetic was using only the
first pixel/element of the input dataset (image or table column) to define
the scatter of its Poisson distribution (which is approximated with a
Gaussian where the background is added to each pixel value). This was a
mistake in the coding!

With this commit, it now properly uses each separate pixel to define the
scatter and loop goes over that also.

This bug was reported by Irene Pintos Castro.

This fixes bug #64138.