MakeCatalog: --std and --sigclip-mean-sb-delta working for clumps
Until now, these two options would produce errors when MakeCatalog was run
with the --clumpscat option (to generate a clump catalog).

With this commit, their respective checks have been added to the
MakeCatalog source, so they can be properly accounted for and the final
catalog can be created.

This bug was reported by Raul Infante-Sainz.

This fixes bug #64146.