MakeCatalog: dashes replaced with underscore in some column names
Until now, the names of some of the MakeCatalog measurement columns used a
dash ('-') instead of an underscore ('_')! However, all the other
multi-word column names contain an underscore. Also, the column name of
'--sigclip-mean-sb-delta' didn't match the option name!

With this commit, the problematic column names have been corrected in the
output table.

This bug was reported by Raul Infante-Sainz.

This fixes bug #64146.