Book: new section explaining the difference between std and error
Until this commit, we did not have a section explaining the difference
between the standard deviation and the error, and sometimes this causes
confusion.

With this commit, a new section has been added under MakeCatalog for this
purpose. Using a practial example, we show the different concepts and how
they can be derived from each other.