ds9-region: properly quoting units to print name columns
Until now, when run in image-mode and given the '--namecol' option, this
script simply printed a zero for the name (ignoring any value!). However,
in WCS-mode, there was no problem and the name was printed.

This happened because DS9's region file format requires a double-quotation
(") for showing the arc-seconds nature of the region size in
WCS-mode. However, in image-mode, it doesn't need anything. As a result, we
would just put the empty string there. However, 'printf' would not "see"
the empty string in image-mode and use the name column instead of units!

With this commit, the problem was solved by placing the 'unit' variable
within double quotations. Therefore, even when it is an empty string,
'printf' will "see" it and put the empty string instead of units. As a
result, it will print the name column in its proper place.

This bug was reported by Zahra Sharbaf.

This fixes bug #64153.