Configuration files: name and value can be separated by =
Until now, only white-space characters could be used to separate an option
name and value in configuration files. But this was inconvenient when you
wanted to use the same configuration in a shell script or a Makefile (which
require a '=' between the name and value).

With this commit, the '=' characters is also accepted as a separator
between option names and values.