Configuration files: --config-prefix=STR allows optional prefix
Until now, if we had several configuration files for multiple instances of
the same program, it was not possible to load them all into the shell
(because same options would overwrite each other's values).

With this commit a new '--config-prefix' common option has been added to
all the programs. It is given a string that can be in the prefix of options
in separate configuration files. This solves the problem above: allowing
complex pipelines to easily load all configuration files into their shell
for later usage of the options.

This completes task #16365.