Query: used value of --dataset in new variable
Until now, we would re-write the user's given value to '--dataset' when it
was given as a summary (for example 'dr3' instead of 'gaiadr3.gaia_source'
in the Gaia database). This re-write would cause a change of pointer, and
thus a core dump when the value was to be written.

With this commit, a new 'datasetuse' variable is defined for the "used"
dataset string, leaving the original user's string to be 'datasetstr'. This
avoids any potential re-write of memory before and after the query's
execution.

In the process, a few minor issues were also fixed:

  - gal_options_parse_name_and_value: a 'break' was missing after each
    'case' statement! This was the original cause of bug #64186.

  - While running Query with Valgrind, I noticed a few memory leaks that
    are also fixed with this commit.

This bug was reported by Zohreh Ghaffari and Sepideh Eskandarlou.

This fixes bug #64186.