astscript-zeropoint: check test of zero point script is added
Until now, 'astscript-zeropoint' did not have a test, when running 'make
check'.

With this commit, all the things for checking the 'astscript-zeropoint'
script is prepared and the check test is done. To do the test it was
necessary to create another test for 'asttable' to make a reference catalog
and star catalog (which are same).

Thanks to Mohammad Akhlaghi for helping me to add this test.