astscript-zeropoint: new --mksrc option to specify Makefile
Until now, when Gnuastro's 'make check' was run on a computer that didn't
already have Gnuastro 0.20 installed (and thus didn't have the
'zeropoint.mk' directory in the necessary location!), the zero point
script's check would fail. This was because we had hard-coded the
installation directory of the Makefile within the script (to be the
finally installed location!).

With this commit, a new debugging option has been added to this script
called '--mksrc'. When this option is given, instead of using the default
installed Makefile, the script will use the file given to this option. By
adding this option to the test script for this script, we were able to fix
the crash mentioned above.

In the process, I also noticed that the order of the installed scripts
within 'bin/script/Makefile.am' was not alphabetical (making it hard to
find the script we want). With this commit the rules to prepare the
installed scripts are alphabetically ordered.

This issue was found with the help of Agata Rożek.