Book: DS9 installation now also works for macOS tarball download
Until now, the 'mv' instructions for SAO DS9 only moved the 'ds9'
executable to '/usr/local/bin'. However, On macOS, the tarball also
includes a 'ds9.zip' to be moved to that directory.

With this commit, the installation instruction starting with 'mv' will now
move both ds9 and ds9.zip into the /usr/local/bin directory.