Book: explanation of DS9 download formats
Until now, the appendix in the book on installing DS9 didn't explain the
CPU architectures that were mentioned on the webpage (which would be hard
to distinguish for someone who is not familiar). Also, when installing
DS9's executables, we would install the tarball also.

With this commit, a short explanation of the different suffixes in DS9's
folders for CPU architectures have been added. Also, the tarball is deleted
before installation.

Furthermore, to make the list of names in 'THANKS' more readable in a
sorted manner, they names in 'THANKS' are now sorted in alphabetic order by
first name, not family name.