astscript-fits-view: print message when input file does not exist
Until now, this installed script wouldn't print any message if the input
file did not exist! The test for the existence was merged with the test for
being a FITS file (with 'astfits'), therefore when it crashed, the script
would not continue.

With this commit, to fix the problem we now explicitly check the existence
of the input before calling 'astfits'.

This bug was reported by Ryan Begley.

This fixes bug #64250.