Query: the new format of Gaia's types for info
Until now, the Gaia servers used strings like "BIGINT" or "REAL" to
identify types for 64-bit signed integers or 32-bit floats. However, some
time in the last few months, this has been changed to the more standard
"long" or "float" strings (similar name changes for other types). As a
result, when Query was run with '--information' on any Gaia dataset, it
would crash because the type was not recognized.

With this commit, after investigating the newly downloaded meta data table,
this change was discovered and Query's type reading functions were
corrected to accound for the new type names also.

This bug was found during a discussion with Rashid Yaaqib.

This fixes bug #64246.