MakeProfiles: point profiles are built in 3D
Until now, when building a point profile on a 3D cube, MakeProfiles would
crash with an error about a dimension with zero length. This happened
because point profiles occupy a single element (pixel/voxel), and required
special attention (as opposed to the rest of the profiles).

With this commit, the necessary check for 3D data has been implemented in
point profiles. Also, I noticed that '--mforflatpix' was not being used for
point profiles, which is un-intuitive (after all a point is "flat"!). So
that was also fixed.

This bug was found by Teet Kuumta.

This fixes bug #64274.