.gitignore: does not ignore gnuastro-internal, only tarballs
Until now, the '.gitignore' file contained a 'gnuastro-*' line. At the
time, it was primarily added for tarballs (with a format of
'gnuastro-X.XX.tar.lz'. However, it would mistakely result in ignoring any
change made in the 'gnuastro-internal' directory!

With this commit, that line has been modified to contain the '.tar'
component to avoid forgetting to commit new files added in the
'lib/gnuastro-internal' directory.

This problem was found while working on the branch with overflow checks
(with Siyang He), and also while working on the error branch with Labeeb
Asari and Fathma Mehnoor. In all these cases, we needed to add new files in
the 'lib/gnuastro-internal/' directory that were not being shown with 'git
status' (and could possibly be forgotten to commit)!