Book: fitting tutorial corrected with --quiet and --fitestimate
Until now, in section of "Least squares fitting", when '--quiet' and
'--fitestimate' were used in an example, the fitted parameters were shown
to be printed on the standard output. While, when two of these options are
used together the fitted parameters are not printed on the standard output.

With this commit, in the example which both of these parameter are used
togther the fitted parameters are not printed in standard output. Some
explanations have also been added in edits by Mohammad Akhlaghi before
merging into the 'master' branch.