Library (fits.h): NaN keys printed as largest value with warning
Until now, when the 'gal_fits_key_write_in_ptr' function was given a
keyword with a NaN value, it would print a warning and let CFITSIO crash
(because a NaN is not supported in the FITS standard). However, it can
sometime happen that an analysis produces a NaN and the keyword needs to be
written.

With this commit, to deal with such cases, this function now writes the
largest possible number for the respective floating point type and will
also print a warning on the terminal, informing the user that this swapping
of values was done.

This bug was reported by Sepideh Eskandarlou.

This fixes bug #64303.