Library (arithmetic.h): new to-1d operator for removing dims
Until now, there was no easy way to remove the dimensionality of an input
dataset (to have a 1D array). But this is sometimes necessary (for example
to do the "flattening" phase of convolutional neural networks.

With this commit, the necessary operator has been added to the arithmetic
library and the documentation contains a fully working example of how to
use it to simulate flattening.

This was suggested by Faezeh Bidjarchian.