Library (txt.c): rest of metadata is read if name is not present
Until now, when the NAME component of Gnuastro's plain-text table format
was not present, the relevant function ('txt_info_from_comment') would
ignore the whole comment line! But this is not what we have described in
the book (we have mentioned that NAME is optional)!

With this commit, the cause of the problem (an artificial 'return'
statement when NAME is empty) has been removed and everything works nicely:
the imported column will have no name, but that is no problem (all the
programs and the table reading functions in the library also accept columns
by their counters).

This fixes bug #64357.