dither-simulate: new script to simulate a dither pattern stack
Until now, there was no easy, high-level way for Gnuastro users to simulate
a certain dither pattern. But this is often necessary when we are planning
an observation.

With this commit, a new script has been added to do the job. This is its
raw format and it will be modified and documented in the next series
commits before this branch gets merged with the 'master' branch.