dither-simulate: analysis done within a Makefile
Until now, the newly added 'bin/script/dither-simulate.in' would simply
create the dithering pattern in series, within the shell script itself.

With this commit, following the experience that was learnt in the Zeropoint
script, a separate Makefile has been written for this script to call. This
significantly improves the speed.