Book: completed first draft of dither-simulate script's documenation
Until now, the documentation in the book for this script only had the
introduction.

With this commit, a first draft of the "Invoking astscript-dither-simulate"
section has been written, describing all the options. Also a new section
has been added to show how we can use this script. Althrough, it is not yet
complete.