Table: new $_all to repeat arithmetic operation on all columns
Until now, if it was necessary to repeat the same operation on all the
columns of a table, users had to repeat the "-c'arith ...'" option many
times. When the number of columns are many, this was annoying (especially
since managing commenting of the single-quotes in a shell loop to later use
in the 'asttable' command is not easy!).

With this commit, a new '$_all' has been defined in Table's column
arithmetic. When an Arithmetic expression contains this string, that
expression will be repeated for all the columns in the table.

This was added after a discussion with Sepideh Eskandarlou.