Book: manipulate all the table columns with one column arithmetic
Until now (in the previous commit), a new feature to do repeated operations
was added in column arithmetic. However, it was not documented!

With this commit, a new box ("cartouche" in Texinfo) is added to the book
to explain how the users can use the new '$_all' feature.

The text was edited by Mohammad Akhlaghi.