Library (fits.h): reading functions also require HDU option name
Until now, when a requested HDU didn't exist, the library would always
print a message suggesting to use '--hdu'. But generally, the problematic
option may be another option (for example '--gridhdu' in Warp!).

With this commit, the FITS reading functions take an extra option for the
caller to optinally give the name of the option that is being used.

This bug was reported by Aaron Watkins.

This fixes bug #64420.