Fixed the errors in tiff_img_write function
Previously, the tiff_img_write function had several errors that
prevented it from functioning correctly. This commit focuses on
addressing these errors and improving the overall functionality
of the function.

With this commit, the following changes have been implemented:
1. Replaced the function TIFFWriteEncodedStrip which was causing
   error with TIFFWriteScanline for writing into tiff file.
2. Updated the corresponding parts of the function to accommodate
   the use of TIFFWriteScanline.
These modifications rectify the errors in the previous commit and
it can now successfully write images to TIFF files as intended.