Library (arithmetic.h): isnotblank works on input with no blanks
Until now, when the 'isnotblank' operator was run on a dataset that didn't
have any blank values, the output would be all zeros (while it should have
been all ones!). This was happening because a separate route was defined
for datasets that don't have any blanks; and in that route, we hadn't
accounted for the flagging of non-blank values!

With this commit, the problem was fixed by setting all the values to 1 when
the input has no blanks.