Book: Minor edits and clarifications
Until now, the default operation of Convolve was not highlighted under each
option which disables one steps (for example '--nokernelflip' did not
mention that by default the kernel is flipped by default). There were also
some typos and non-clear statements in the comments of some library
functions.

With this commit, a link to the section describing the operation of
Convolve has been added under those options. Furthermore, the typos and
clarifying points have been implemented in the comments within the library
source code.