Library (arithmetic.h): new trim operator
Until now, there was no easy way to trim blank rows and columns in an
image. But this is an oftenly necessary operator, especially within the
context of the dithering pattern that we are adding in this branch.

With this commit, a first draft implementation of this operator has been
added and will be tested more during the work on this branch.

Also, with this commit:
- The process of writing a tutorial for the dither simulation installed
  script has begun.
- Furthermore, the tutorial on re-gridding different filters into the same
  grid for producing a color image has also been moved to the Tutorials
  chapter (previously it was under the ConvertType program, making it hard
  to find for a new user).

- The reminder parts of the book that didn't follow the
  one-sentence-per-line convention have been corrected.