Book: first draft of new dither tutorial, upaded NEWS for it
Until now, the tutorial showing the logic and methods of using the newly
added 'astscript-dither-simulate' script was incomplete. Also, the 'NEWS'
file had not been updated with some of the newly added features in this
development branch.

With this commit, the necessary steps have been added to complete a first
version of what I had in mind for it. It still needs to be further tested
and edited of course; but this sufficiently complete for merging into
'master'. Also, the 'NEWS' file has been updated.