Installed scripts: sources are renamed to .sh instead of .in
Until now, the source of the installed scripts were files that are named as
'*.in'. However, this suffix is mostly used by tools like Automake to make
temporary intermediate files (like the 'Makefile.in' files). As a result,
line-counting software (like CLOC) ignore the '*.in' files and this is
harmful for general statistics of Gnuastro's source code.

With this commit, to solve the problem, the source file of the installed
scripts ends in '.sh' (as a shell script should be).