Library (wcs.h): gal_wcs_to_cd accounts for altlin==3
Until now, this function would only consider scenarios where the input file
had a single rotation matrix (in particular, 'CDi_j' and 'PCi_j'). However,
it can happen that an input FITS file has both! In this case, it would
abort with a bug notice.

With this commit, when both representations are present, it will check if
the values are the same or not. If they are, then it will just re-write
'altlin' so WCSLIB uses the 'CDi_j' representation (as requested from this
function). But if they are not the same, then it will abort with an
informative error to let the user know about the problem and how they can
fix it.