Table: new options of --info-num-cols and --info-num-rows
Until now, there was no easy way for the user to easily print the number of
rows or columns in a table. They would either have to read the FITS
keywords (which would only work on FITS tables) or try parsing the output
of '--information' (which would be buggy and non-trivial).

With this commit, the two new options above have been added to Table to
simplify extracting these two numbers (that are often necessary in
scripts).