Book: Zeropoint tutorial brought into the Tutorials chapter
Until now, the tutorial describing how to obtain the zero point of an image
was located within the documentation of 'astscript-zeropoint'. This made it
hard to locate, and separate from the other tutorials.

With this commit, it has been brought up to the "Tutorials" chapter with
all the rest of the tutorials.