Library (arithmetic.h): mknoise-poisson real Poisson distribution
Until now, the 'mknoise-poisson' operator of Arithmetic would just make a
Gaussian approximation of the Poisson distribution: it would not be
integers, and it could include negative values.

With this commit, we are now using the GNU Scientific Library (GSL)'s
Poisson distribution random number generator to produce correct Poisson
noise. To keep the old behavior, a new 'mknoise-sigma-from-noise' operator
has been added and the description of 'mknoise-sigma',
'mknoise-sigma-from-noise' and 'mknoise-poisson' now includes hands-on
examples for the readers to clearly identify when each should be used.

Furthermore, some minor modications were made in the book:

 - Moire pattern tutorial has been moved after the dither pattern tutorial.

 - The dither simulation script has been mentioned in the list of
   Gnuastro's programs of the book as well as the 'README' file.