MakeNoise: deleted from Gnuastro
Until now, we had two ways of adding noise to an image: the MakeNoise
program or the 'mknoise-*' operators of Arithmetic. This was done
intentionally to gradually depreciate MakeNoise as a separate program
because the Arithmetic operators are much more modular and customizable.

With this commit, MakeNoise has finally been removed since the operators do
all that the program did. The "Noise basics" sub-section of the book has
also been moved under the Arithmetic program.

In the process, some of the 'Makefiles.am' files have been cleaned to have
one-line per file name in the variable definitions to help in readability.