MakeCatalog and Segment: first HDU name changed to INPUT-NO-SKY
Until now, the name of the first extension of Segment's output was 'INPUT'.
But this does not convey that by default Segment expects the sky to be
already subtracted (if not, the '--sky' should be used).

On the other hand, until now, in MakeCatalog the name of default values HDU
('--valueshdu') is a number ('1'). However, the default value for the other
input HDUs are clear names (guiding the user on what is expected). This was
because until now MakeCatalog assumed the first extension is values
file. While in some cases maybe the first extension should not be used for
values, therefore the result that will obtained from MakeCatalog are going
to be wrong (without any descriptive error).

With this commit, first of all the name of first extension of Segment's
output is changed from 'INPUT' to 'INPUT-NO-SKY'. Finally, the default
values HDU in MakeCatalog has changed to a descriptive name
('INPUT-NO-SKY').