Segment: --keepmaxnearriver has an effect on the output
Until now, when the user gave '--keepmaxnearriver' the number of output
clumps would be the same as before, even when clumps existed that had a
maximum that was touching a river. This happened because when checking for
neighboring river pixels, Segment was looking for 0-valued pixels, but they
will be negative!

With this commit, the problem has been solved by properly checking the
neighboring pixels of the maximum pixel as those that are less than 1.

This fixes bug #64541.