Arithmetic: tofilefree- operator correctly deals with directories
Until now, when the 'tofilefree-' operator of Arithmetic was given a file
name that included a directory, it would crash. This was because the
proper length of the string was not been checked.

With this commit, the problem is solved by using the length
'OPERATOR_PREFIX_LENGTH_TOFILEFREE' instead of the mistaken
'OPERATOR_PREFIX_LENGTH_TOFILE' length!

This fixes bug #64544.