Query: use libcurl Library for Remote Server Access
Until now, the Query program has utilized the 'curl' executable from
the 'cURL' package for remote server interactions. However, this approach
involves a 'system' call, which limits our control over outcomes.

With this commit, we transition to using the 'libcurl' library, to access
remote repositories. Unlike the 'curl' executable, 'libcurl' offers a
more controlled approach that keeps the execution within Gnuastro's
programs. Now, it first checks if the 'libcurl' library is availiable on
the system. If it's present, the 'libcurl' library is employed to retrieve
data. Otherwise, the data is retrieved using 'curl' executable as before.