Configuration: HAVE_LIBCURL passed onto the source
Until now, the source of Gnuastro had no indication (through a standard
'HAVE_LIB*' macro) if the libcurl library was found at configuration time
or not. But this was necessary for the work done on using libcurl within
Query.

With this commit, the necessary steps have been added in the configure
script and 'HAVE_LIBCURL' is now available within the source of Gnuastro's
library and programs.