psf-stamp: mask the extra objects after applying the threshold
Until now, for masking the extra objects, first of all we masked the
objects (OBJECTS of segment output) that have different values with the
main object. Then in the object of the main star we just kept clumps
(CLUMPS of the segment output) of the main star and then masked the other
clumps. But when the user requested a S/N threshold some pixels with high
value were not masked.

With this commit, after applying the threshold we masked the extra regions
that remain and do not touch the main star.