MakeCatalog: same input clump labs when NUMLABS not in clumps image
Until now, when the 'NUMLABS' keyword was not in the clumps input of the
MakeCatalog, it would reset the labels to start from one. This was very
inconvenient and required the user to map the new labels with the original
labels to use them.

With this commit, the internal mapping is kept within MakeCatalog and is
used to put the original labels in the clumps afterwards.