Installed scripts: unified error message formats
Until now, some of the error messages in the installed scripts started with
'ERROR:' after the script script name. But this is not a convention within
Gnuastro's programs (or in the GNU Coding Standards in general). Also, when
checking the '--*center' option value, some of the scripts did not print
the original string! Further more, one of the error messages ended with a
'.' (which is against the GNU Coding Standards).

With this commit, all the problems above have been fixed.