Book: suggestion to use the -s option of seq with many columns
Until now, the '_all' option has proven invaluable for appending columns at
either end of a table, situations arise where the task is to selectively
delete or retain specific columns within a table, particularly when the
table contains numerous columns. In such cases, generating a sequence of
the desired columns becomes necessary.

With this commit, an explanation has been added into the Gnuastro book for
such cases. This addition outlines the methodology to create a sequence
encompassing an arbitrary sequence of column numbers using the 'seq'
command. By utilizing the '-s' flag, a comma delimiter is effectively
inserted between each column number, rendering it suitable for the '-c'
option in the Table program.