doc/release-checklist.txt: minor edits
Until now, given the recent changes in the Gnulib's
build-aux/announce-gen', the release checklist had the version number
'0.20' for the previous release! But This will not be valid in future
releases. Also, the directory to run the commands was not explicitly
mentioned!

With this commit, both issues have been implemented.