doc/release-checklist.txt: adding steps for conda
Until now, we would rely on a stable release to have a Conda check. But in
case there is a bug, the only solution would be to make a new stable
release (which is very inconvenient and occurred in version 0.16!).

With this commit, a set of steps has been added in
'doc/release-checklist.txt' to do a fake Conda build to find any potential
bugs before an alpha or stable release.