Fits: --copykeys finds exact match when given a name
Until now, when two keywords had the same first characters (for example
'SBLMAGPX' and 'SBLMAG' of MakeCatalog's outputs), and the longer one was
first, if the user gave '--copykeys=SBLMAG', the longer one would be
copied, not the exact match! This happened because we were just checking
the first set of characters.

With this commit, when the first set of characters match, we also check the
next character and only consider it a match when the next one is a SPACE
character or an '=' (showing that the keyword name has finished according
to the FITS standard).

This fixes bug #64610.