Library (polygon.h): new function to find polygon area on sky
Until now, the only available function to find the area of a pixel was the
'gal_polygon_area' function, which assumed a flat coordinate
system. However, this function was used in the 'gal_warp_pixelarea'
function (that calculates the pixel are on the spherical sky
coordinates!). As a result, the reported areas were wrong (would change
with declination).

With this commit, that function has been renamed to 'gal_polygon_area_flat'
to remind the callers that it is for a flat coordinate system. For
spherical coordinate system, a new 'gal_polygon_area_sky' has been
added.

Also, since the area of the pixels are independent of the position on the
sky (CRVALi in the FITS keywords), the 'gal_warp_pixelarea' now does the
WCS conversion with a temporary CRVALi of (180,0). To do this, it was
necessary to define a new function to change CRVAL (because WCSLIB keeps
and uses low-level internal parmaters that will not change when we simply
change 'wcs->crval').

This fixes bug #64615.