Table: new eq-j2000-on-flat operator for plotting RA and Dec
Until now, Gnuastro didn't have an easy way to plot RA and Dec (while
accounting for the curvature of the sky; which can be significant when the
distances go larger than a degree or so).

With this commit, a new column arithmetic operator called
'eq-j2000-on-flat' has been added to do this on a given RA and Dec
column. Allowing users to easily plot the points with any plotting tool
they prefer.