Table: new eq-j2000-from-flat operator
Until now, Table's column arithmetic only had the 'eq-j2000-on-flat'
operator for converting spherical to flat coordinates. But the inverse of
this operator is also commonly necessary. and the 'on' was ambiguous: was
it "from" or "to".

With this commit, this operator has been renamed to 'eq-j2000-to-flat' and
the newly added inverse operator is now called 'eq-j2000-from-flat'. This
new operator is now also used in the dither tutorial and because the
tutorial had become long, it is now broken into several sub-sections.