Library (arithmetic.c): error for binary integers of same width
Until now, when both operands of binary operators in the arithmetic library
had the same width, but different signs, a warning was printed that the
output might be wrong. But, when the output may be wrong we need the Error,
not warning. Because the main purpose of each function is to produce output
correctly, not just to create output. Also, there were some typos within
the comments of various parts of the source.

With this commit, this warning has been changed to an error. Besides, the
typos have been fixed.

This issue was originally raised by Sepideh Eskandarlou.