psf-stamp: keep saturated pixels after applying the SN threshold
Until now, when the '--snthresh' option was applied into the saturated
stars the output was completely NAN. This was because of the central pixels
being NaN, so the extracted labels (of the center) became 0 and the whole
image would become NaN.

With this commit, the saturated area is filled with a non-existant value
(maximum + 1) and then the threshold is applied. After all other regions
above the threshold have been removed, the NaN values are returned.

This fixes bug #64635.