MakeCatalog: value of --upmaskhdu correctly printed on stdout
Until now, MakeCatalog incorrectly printed the value to the '--hdu' option
for the statement it printed on standard output. However, for the actual
analysis, the correct HDU (from the correct file) was used. This was
confusing!

With this commit, the correct value of '--upmaskhud' is printed on standard
output.

This bug was reported by S. Zahra Hosseini Shahisavandi.

This fixes bug #64643.