Library (wcs.h): convert points from one coordsys to another
Until now, we had no function to convert a set of coordinates in one
celestial coordinate system to another (for example to convert the RA and
Dec columns in a table to galactic coordinates).

With this commit, the necessary functions have been added within the WCS
library; enabling 30 new operators for Arithmetic (for conversion to and
from all the 6 exisiting coordinate systems).