Book: corrected demos of gal_fits_img_read_to_type
Until now, the 'gal_fits_img_read_to_type' function examples didn't include
the recently added 'hdu_option_name'. This function has been used in some
of the programs in the library demo and it was necessary to correct this.

With this commit, this parameter has been added to this function available
in the library demo programs. Furthermore, some incorrectly written
"deconvolution" words were corrected in other parts of the text.