pointing-simulate: new name for the dither-simulate script
Until now, we used the term "dither" instead of the more generic
"pointing". This could be confusing because a "dither" is usually used for
small displacements on the sky and can be confusing when the observing
strategy has to account for large spacing (and needs to re-focus the camera
and etc).

With this commit, the name of the installed script is now set to
'astscript-pointing-simulate', also the text of the book has been edited to
clarify these definitions.