developer-build: new option to specify compression
Until now, the developer build script would only produce a 'tar.lz'
compressed tarball. However, in some cases, it is necessary to produce a
different compression.

With this commit, a new '--compression' option has bene added to the
'./developer-build' script, allowing users to select either '.lz' or '.xz'
compressed tarballs. If any future format is necessary, it is easy to add.