doc/release-checklist.txt: improved/edited conda-forge instructions
Until now, the checklist to confirm the build of Gnuastro on CondaForge
wasn't actually tested and didn't have all the necessary details.

With this commit, after an application of all the steps in a forked
repository and confirming the successful build on all CondaForge
architecture in Gnuastro, the text has been edited and made more clear.

Also, a small typo was corrected in the 'developer-build' script for the
name of the default compression to use.

The instructions to do a test build on CondaForge were initially provided
by Sebastian Luna-Valero; thanks a lot Sebastian!