NEWS: program names (sub-titles) in org-mode format
Until now, while the top-level (release) and sub-sections (new features, or
bug fixes) were in org-mode format (starting with '*' and '**'), the names
of the programs had no special formatting! This would make it hardly useful
when looking at the 'NEWS' file in org-mode.

With this commit, program names of the top four releases of Gnuastro are
now also in org-mode sub-section format (starting with '***').