Book: typo corrected in installation from package managers
Until now, a '\' character was missing for one of the of the long commands
(that was broken into multiple lines). As a result, a raw copy-paste of the
command would cause inconveniences for the users (the third line would be
interpretted as an independent command).

With this commit, the typo has been corrected.

This typo was reported by Michael Stein.