Book: example in commit guide lines now fits in PDF page width
Until now, the example of the commit body (within the commit guidelines
section) did not fit with the page size and it was out of the normal book
page size. Because instead of writing 75 characters in each line, a
sentence was written in each line.

With this commit, the example commit message has been adjusted to have a
fixed line width of at most 73 characters (75 would also go beyond the text
limits in the PDF).