Library (convolve.h): new argument for convolution over blanks
Until now, the library's convolution functions would always ignore blank
pixels. In most scenarios this is fine and expected, but in some
situations, users may want to use convolution as a proxy for interpolation
or dilation.

With this commit, a new 'conv_on_blank' argument has been added to the core
convolution library to allow for this feature when the user wants this
behavior. Also, the Convolve program now has a new '--conv-on-blank' option
for users to directly activate this feature on the command-line.

This was suggested by Raul Infante-Sainz.