Webpage: citations updated to include the Gnuastro ADS library
Until now, for a listing of the citations, we were only pointing to the
NoiseChisel paper. But the number of Gnuastro papers is increasing and
people sometimes use the ASCL or other papers only. Therefore, I recently
defined a public ADS library to collect Gnuastro-related papers and thus
their citations.

With this commit, the ADS library link has been used in the top webpage.