astscript-rgb-image: Adding script for generating color image
Until this commit, we had a script that uses Gnuastro for generating color
images in another repository. This script is now mature enough to be
included as part of Gnuastro as a script. It is based in Lupton et al.
(2004): http://doi.org/10.1086/382245. In short: it takes three images
(rgb; red, green, blue) to create a color image. The pixel value histograms
are manipulated appropriately in order to show the low and high pixel
values using the asinh function. The high pixel values are shown in color,
while the low pixel values are show in black and/or gray.

With this commit, the script has been added to Gnuastro as it is at the
moment. From now, necessary changes as well as documentation and
information will be added.