astscript-rgb-image: select black or gray background
Until this commit, the rgb-image script generates two output images. One
image with color and black background, and another one with color and gray
background. While this is good because both images are obtained at once, it
is not good in other scenarios like in pipelines or just if one of the two
images is not wanted. In addition to that, a predefined output name for the
gray-background image was selected. This is not good.

With this commit, a new option has been added: --grayback. This option
allows the user to ask for the gray-background image and not generate the
black-background image. As a consequence, time is saved. Also, now the
script can be easily used in pipelines where systematic inputs and output
names are required.